Rename `git_config_entry` module to `remote`
The `GitConfigEntry` enum has been removed from the module and now only
contains the `GitRemoteRepo` type.
